refactor: 관심사 분리
useReAuth 훅 제거후 App.tsx에 useEffect 작성
