+++ src/App.tsx
@@ -1,11 +1,23 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import Router from './Router/Router';
 import { DefaultTemplate } from '@/components';
-import useReAuth from './hooks/useReAuth';
+import { useAppDispatch } from './store';
+import { getItemFromStorage } from '@/utils/storage';
+import { initiateAuth } from './store/userSlice';
 
 const App = () => {
-  useReAuth();
+  const dispatch = useAppDispatch();
+
+  useEffect(() => {
+    const accessToken = getItemFromStorage('accessToken');
+
+    if (!accessToken) {
+      return;
+    }
+
+    dispatch(initiateAuth());
+  }, [dispatch]);
 
   return (
     <BrowserRouter>
